requirements.txt
Django==4.2.3
django-cors-headers==4.1.0
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==20.1.0
Pillow==10.0.0
psycopg2-binary==2.9.6
requests==2.31.0

pandas~=2.0.3